fbpreprocessor: reset rate limit on login/logout
fbpreprocessord restricts the number of pseudonymizations to 5 per 30
minutes. If a user generates a lot of firmware dumps, the next user
might be blocked from processing feedback reports for up to 30 minutes.

Instead, we reset the state of the rate-limiting logic when the user
logs in or out.

BUG=b:323601976
TEST=cros-workon --board=${BOARD} start fbpreprocessord
TEST=FEATURES=test emerge-${BOARD} fbpreprocessord
TEST=cros deploy ${DUT} fbpreprocessord
TEST=for i in `seq 1 10`; do echo 1 > /sys/kernel/debug/iwlwifi/0000\:00\:14.3/iwlmvm/fw_dbg_collect; sleep 5 ; done
-> no firmware dumps created after 5 first ones
TEST=logout, log back in and re-run the previous line -> 5 firmware
dumps are created

Change-Id: I26d893004bbab5a01f74b25100a9d941ebe398b3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5264227
Reviewed-by: Manish Mandlik <mmandlik@chromium.org>
Tested-by: Nicolas Norvez <norvez@chromium.org>
Commit-Queue: Nicolas Norvez <norvez@chromium.org>