shill: Use cached service list in Manager::AutoConnect()
Under the current architecture in shill, it's hard to guarantee that
service list or a service itself is unchanged when calling
Service::AutoConnect on a service. Copy the service list with weak
pointers do the loop to be safer.

There shouldn't be any functional changes introduced by this patch.

BUG=b:323386660
TEST=FEATURES="test" emerge-$BOARD shill

Change-Id: Ic888dc64ab7f29f2d31cedf28c0425d094df45cd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5264657
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>