lorgnette: Set protocol_type for discovered scanners
The protocol_type field contains a general description of the
protocol/backend used to access a scanner, such as Mopria, WSD, or
epson2.  Since we don't know the actual name of the protocol implemented
by most of the backends, set this field to the backend name in all cases
except the procotols implemented by sane-airscan.

BUG=b:322325211
TEST=lorgnette_cli discover --show_details

Change-Id: Id2aa964c16381c6b31497cb6882d425a97293907
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5277917
Reviewed-by: Nathan Muggli <nmuggli@google.com>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>