init: Add disable_lvm_install mode for clobber
For testing migration between ext4 and LVM stateful partition,
we need clobber to reset the partition to ext4. Add an extra
arg that would allow us to format the stateful with ext4.

Additionally, add a USE flag for making this the default
behavior. For boards that are migrating to LVM, this will
be default on (as an escape hatch in-case the migration puts the
device in an undesirable state).

BUG=b:308692022
TEST=CQ

Cq-Depend: chromium:5532550
Change-Id: Ib4b5f9f63d0a6c139b8bcc5c3eba17cc1f5ebdc0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5271568
Reviewed-by: Tim Bain <tbain@google.com>
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Commit-Queue: Sarthak Kukreti <sarthakkukreti@google.com>
Tested-by: Sarthak Kukreti <sarthakkukreti@google.com>