shill: PortalDetector uses OnceCallback to represent the running state
Originally, PortalDetector uses a boolean to represent whether it's
running or not, and uses a RepeatingCallback for the callback.

This CL changes the RepeatingCallback to OnceCallback. With this change,
we could reduce one redundant variable. Also, OnceCallback shows the
one-to-one mapping relationship between the Start() method and the
callback.

BUG=b:305129516
TEST=FEATURES=test emerge-brya shill

Change-Id: I208749bb20325c4fa978f636108ddd120a56f199
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5272001
Commit-Queue: ChromeOS Auto Retry <chromeos-auto-retry@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>