shill: cellular: StopModem won't return WrongState
The WrongState error would have been expected when trying to disable
cellular and the modem was in Failed state (i.e. without SIM).

ModemManager now allows the low power mode request even if the modem
is in Failed state, so there is no reason to handle this error
differently, as it should not happen again.

BUG=b:315750677
TEST=FEATURES="test" emerge-$BOARD chromeos-base/shill

Change-Id: I2ecbd1557e68b2898e2d11d5ba5a416be3a81a66
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5272542
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Aleksander Morgado <aleksandermj@google.com>
Reviewed-by: Rukun Mao <rmao@google.com>
Tested-by: Aleksander Morgado <aleksandermj@google.com>