init: Turn 'pre-shutdown' into a task
As documented in $BUG, the 'pre-shutdown' job currently allows jobs like
'boot-services' to stop in parallel with the 'reboot' job. The sequence:

1. send reboot signal to upstart
2. 'reboot' job move to starting
3. #2 triggers 'pre-shutdown' to starting
4. 'pre-shutdown' doesn't have anything to do -> 'started'
5. #4 allows 'reboot' to continue
6. pre-shutdown moves to 'stopping'
7. #6 triggers 'boot-services' to stopping
8. when all triggers in #7 are complete, 'pre-shutdown' moves to
   "stopped"

Note that #5 and #7 aren't synchronized; 'reboot' runs in parallel with
tearing down 'boot-services' (and anything that stops along with it).
This yields unexpected race conditions, and causes 'process_killer'
(part of 'reboot') to have to kill excess jobs.

If we turn 'pre-shutdown' into a task, then (per init(5)):

  "the act of starting the job is not considered to be finished until
   the job itself has been run and stopped again"

That is, #5 moves after #8 (and therefore #7), and we elide the race.

BUG=b:267224195
TEST=`initctl log-priority debug; reboot`
TEST=check reboot logs for fewer 'process_killer' events

Change-Id: I7c581e23d99a2633c53dfc37aee7cd734ab31050
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5367740
Reviewed-by: Sarthak Kukreti <sarthakkukreti@google.com>
Tested-by: Brian Norris <briannorris@chromium.org>
Commit-Queue: Brian Norris <briannorris@chromium.org>