hiberman: Check available memory after dropping the page cache
CL:5226561 ("hiberman: Get rid of the giant allocation") removed a
flawed check of whether the system has enough memory to hibernate,
and introduced another check with a similar purpose. The intention
was to perform the check after dropping the page cache (even the
commit message says that), but in the CL that landed the check is
performed before the page cache is dropped. As a result it can
happen that a hibernation is aborted because supposedly there is
not enough memory available for storing the hibernate image, when
there might be enough memory after dropping the page cache.

Perform the check after dropping the page cache, as initially
intended.

BUG=b:324115636
TEST=open tabs and Android apps
  hibernate and resume on brya

Change-Id: Ic26b2c430f30226396337a30189a414375afb672
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5272869
Tested-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Brian Geffon <bgeffon@chromium.org>
Commit-Queue: Matthias Kaehlcke <mka@chromium.org>