shill: wifi: configure pairwise and group ciphers
We need this for WiFi 7. Note that we pass these unconditionally because
crrev/c/5272589 checks for driver support, so we don't need to be
concerned with passing unsupported pairwise and group ciphers.

BUG=b:308681208
TEST=Ran SimpleConnect.wpa3{mixed} on AC7265, BE200, and Marvell with
GCMP pairwise ciphers, everything was able to connect (BE200 with GCMP
and others with CCMP due to lack of GCMP support).

Cq-Depend: chromium:5272589
Change-Id: I3e94279e4b840e44c4c6dc1ee885d61f214d0227
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5273188
Reviewed-by: Jintao Lin <jintaolin@chromium.org>
Reviewed-by: Kevin Lund <kglund@google.com>
Tested-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Kaidong Wang <kaidong@chromium.org>
Commit-Queue: Matthew Wang <matthewmwang@chromium.org>