camera: tools: mctk: Fix nullopt use in remap code
In V4lMcRemap::CreateFromYamlNode(), name and name_regex may be
std::nullopt, yet are used when constructing an V4lMcRemapEntry,
crashing the program.

What actually needs to happen is that empty strings are passed for the
values that are nullopt - this patch fixes this.

BUG=b:324018288
TEST=mctk --load-device /dev/media0 --dump-yaml test.yaml
     mctk --load-device /dev/media0 --merge-yaml test.yaml

Change-Id: Id2434506582937e292ca5ee5576423322a3441eb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5276938
Tested-by: Max Staudt <mstaudt@chromium.org>
Reviewed-by: Hidenori Kobayashi <hidenorik@chromium.org>
Commit-Queue: Hidenori Kobayashi <hidenorik@chromium.org>