cryptohome: Extract even more UDA APIs
This relocates more of the objects that UDA initializes into the
SystemApi+BackingApi constructs. I believe this extracts most of what's
easily constructable, because most of the (unfortunately many) remaining
objects are constructed asynchronously and so can't really be created up
front using the same approach used here.

I had to add some complexity to the mock system APIs to make this work
because we needed the choice of the precise mock setup (specifically,
what objects are mocked and what are real) to be configurable in tests.

I also ditched the ability to configure the niceness of the mocks
because we're only using NiceMock and making it configurable made the
code much, much messier (I needed a template template template
parameter, which is usually a bad sign).

BUG=none
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I407ba1db232f50e6f13c9dbebcd3cda9f33b15db
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5277919
Reviewed-by: Zi Lin <lziest@google.com>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>