login: Replace OpenBrowserProcFile with ReadBrowserProcFile
Instead of providing a helper that opens a file in the browser's pid
directory implement one that reads the contents and returns a string.
This prevents some code duplication, which will be especially helpful
with the upcoming logging improvements, some of which will rely on the
proc files.

BUG=b:324016659
TEST=Trigger a browser hang and check logs

Change-Id: I64e1d7c74ce9eb638c3bb4850cb1e508026c4b36
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5291379
Tested-by: Kornel Dulęba <korneld@google.com>
Commit-Queue: Kornel Dulęba <korneld@google.com>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>