patchpanel: do no reset iptables state when shutting down
This patch changes the exit sequence of patchpanel to skip any cleanup
operation in iptables (and ip6tables). When the device itself is
shutting down, resetting the iptables state is not necessary. And when
patchpanel has to restart, the start sequence already assumes the worst
and tries to reset the iptables state upfront.

BUG=b:324130592
TEST=FEATURES=test emerge-$BOARD patchpanel

Change-Id: Iae79781233743453585592357b83be07da1e2aae
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5273752
Tested-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>