shill: validate the network after the remaining time is over
The CAPPORT API might hint the remaining time until the portal is
closed. This CL makes the NetworkMonitor trigger the network validation
again after the remaining time is over.

Also, this CL clears the fields of remaining time and remaining bytes
when the portal is captive, because these fields are only meaningful
when the portal is open.

BUG=b:305129516
TEST=FEATURES=test emerge-brya shill

Change-Id: I24038fbc90dcf67f7eeffb867beae114de6ae35b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5276947
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Tested-by: Chih-Yu Huang <akahuang@chromium.org>