camera: fix setup to run over ANGLE
Add a camera GL loader library.

This library goal is to select to correct GL library depending on the
angle state defined by the file `/run/camera/angle_disabled`.

It adds a listener in the cros-camera app to update the angle state in
the filesystem and restart the service if the chrome flag has changed.

Only the functions used by the cros_camera_service are exposed through
the camera GL loader library.

BUG=b:277798516
TEST=run cros_camera_service on top of ANGLE and NativeGL

Cq-Depend: chromium:5317583
Change-Id: I9c46e8b497485e348c536acd001819934e00dd90
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5318752
Reviewed-by: Prahlad Kilambi <prahladk@google.com>
Commit-Queue: Romaric Jodin <rjodin@chromium.org>
Tested-by: Romaric Jodin <rjodin@chromium.org>
Reviewed-by: Ricky Liang <jcliang@chromium.org>