update_engine: move programs & clients behind USE=-cros_host
Normally this isn't needed because the targets aren't depended on by
the main group(all) target, but it looks like the pkg-config logic
is always run & expanded regardless of the target being built.  This
causes the build to need the various .pc files to be installed even
if the headers & libs aren't used.  To mitigate this, we need to put
the non-SDK targets behind a `!use.cros_host` check.

BUG=b:299981037
TEST=CQ passes

Cq-Depend: chromium:5282634
Change-Id: Iff64a35efb0c560b1c1f2b46bfe8189c54bd32a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5279702
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>