shill: tethering: Refresh tethering capabilities on Modem info change
Refresh the tethering capabilities property when the cellular modem
information, such as modem type and FW version is updated.
Currently, the property is only modified when a client reads the
property.

BUG=b:324310282
TEST=On redrix, verify Hotspot is displayed after rebooting
when using the experimental flag. Use dbus commands to set/unset the
experimental flag and verify the UI is updated immediately.
Monitor dbus messages to verify the capabilities are updated
accordingly.

Change-Id: I6bb027f271d047b60d5edd3fe856f8a8d1ab007e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5281651
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Jintao Lin <jintaolin@chromium.org>
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Andrew Lassalle <andrewlassalle@chromium.org>