vm_tools: concierge: avoid passing VmInfo when sending "starting up" signal
When sending "starting up" signal the VM is not fully constructed, so
it is not clear what data is being broadcast. Additionally we should
strive to use protobuf data only at the perimeter of the service, when
returning a result in response to a method call or sending a DBus signal.

Rework SendVmStartingUpSignal() to accept the data items that will be
transmitted as arguments.

BUG=b:315302937
TEST=CQ

Change-Id: I7e3d7589d42c4e939e924ea5366afd0c436c00cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5288322
Commit-Queue: Dmitry Torokhov <dtor@chromium.org>
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Tested-by: Dmitry Torokhov <dtor@chromium.org>