vm_tools: concierge: introduce ToVmInfo() and ToVmStatus() helpers
There are multiple places in concierge where VmInfo protobuf structure
is being constructed. The initialization is spread around, and it is
not clear what data is being copies over.

It is possible to construct the VmInfo data from VmBaseImpl::Info
which can be obtained for a VM by calling vm->GetInfo(). Intruduce
a pair of helpers ToVmInfo() and ToVmStatus() and use them throughout
the code.

BUG=b:315302937
TEST=CQ

Change-Id: I714a777729f0a34640a860ed13551120e2e07ec0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5288323
Reviewed-by: Junichi Uekawa <uekawa@chromium.org>
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Commit-Queue: Dmitry Torokhov <dtor@chromium.org>
Tested-by: Dmitry Torokhov <dtor@chromium.org>