vm_tools: concierge: avoid passing around VmInfo when sending "started" signal
Protobuf version of VmInfo should only be used at the perimeter of the
service, when returning a result in response to a method call or
sending a DBus signal.

Rework Service::HandleControlSocketReady() and its callees to operate
on VmId and similar internal data types.

BUG=b:315302937
TEST=CQ

Change-Id: I62f5882781505130730ad955ec10956518b61a62
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5288324
Tested-by: Dmitry Torokhov <dtor@chromium.org>
Reviewed-by: Junichi Uekawa <uekawa@chromium.org>
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Commit-Queue: Dmitry Torokhov <dtor@chromium.org>