verity: add local blake2b implementation
The implementation is taken from official reference repository
https://github.com/BLAKE2/BLAKE2/tree/master and trim down a bit
with no code modifications otherwise. Lint complains a lot about
that code, but we would never need to modify it, so it should not
be a problem. Some other checks are being shunted in PRESUBMIT.cfg

BUG=b:322886034
TEST=unittest

Change-Id: I726c71779c32f662597dd703eed53956fb80e87f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5289835
Tested-by: Daniil Lunev <dlunev@chromium.org>
Commit-Queue: Daniil Lunev <dlunev@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>