resourced: Introduce Cpuset responsible for cpuset format.
This is a refactoring in cpu_utils.rs. The original implementation was
not efficient especially at HotplugCpuAction::OfflineSmallCore path
which generates cpuset string from a list of u64 and parses it again.

This expands the cpus in range format. However the overhead should be
acceptable because chromebooks should not have too many cpu cores.

BUG=b:324362842
TEST=cargo test

Change-Id: I8330a0172ade41c469affedd2ace60361b26b5c4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5289141
Commit-Queue: Shin Kawamura <kawasin@google.com>
Reviewed-by: Hsin-Yi Wang <hsinyi@chromium.org>
Tested-by: Shin Kawamura <kawasin@google.com>
Reviewed-by: David Stevens <stevensd@chromium.org>