shill: Report shill_id in P2P commands
P2P API commands do not correctly report the shill_id of the P2P GO and
Client devices that are created from the commands. The fix is as simple
as including the shill_id in the call to PostResult in P2PManager

Also add a check to the unit tests that would have caught this issue.

BUG=b:324641607
TEST=anahera-rev4 ~ # dbus-send --system --fixed --print-reply --dest=org.chromium.flimflam / org.chromium.flimflam.Manager.CreateP2PGroup dict:string:variant:
/0/result_code success
/1/shill_id 0
TEST=FEATURES=test emerge-brya shill
Fails without fix, passes with fix.

Change-Id: I26b66cb430f5de99c712a9b76947e6dce97c8ef6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5292888
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Jacek Siuda <jsiuda@google.com>
Tested-by: Kevin Lund <kglund@google.com>
Commit-Queue: Kevin Lund <kglund@google.com>