shill: Increase nf_conntrack_expect_max
Increase the value from 1024 to 4096 so that it can hold more entries.
The expect table is mainly used by the SSDP helper in conntrack. The
general network functionality won't be affected even if this table is
full, but there will be a log of warning logs in messages. The main
purpose of this change is to reduce the number of logs in the lab
environment.

BUG=b:296805772
TEST=Changed the size on a lab machine, checked that there were no more
  error logs.

Change-Id: Ibea20c6966936c55c2b0491ae7c51b0802e1424c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5294994
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Auto-Submit: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Jason Iman <jasongustaman@chromium.org>