shill: network: fetch traffic counter with the primary network
This patch makes Device uses the interface name of its primary Network
for fetching the traffic counters from patchpanel.

This allows to get the correct traffic counter values for the default
PDN connection on a Cellular network. Note that tethering traffic on a
secondary DUN network is still ignored and not added to the Cellular
Service's traffic counters. For other physical networks, there is no
difference between using the Device's link name and the primary
Network's interface name, they are always the same.

The initial implementation for the aggregation of raw traffic counters
relied on the Device to fetch the traffic counters from patchpanel. When
multiplexing was enabled on Cellular networks in m117, the network
interface used for the actual PDN connection and the underlying network
interface of the Device became separate. This broke traffic counter
aggregation for Cellular Service, even though the underlying raw traffic
counters were still functional.

BUG=b:273741930
BUG=b:314018558
TEST=FEATURES=test emerge-$BOARD shill. Flashed jinlon (hatch) and rebooted
the DUT, checked that shill now gets the traffic counters of the default
PDN network as expected (tested with Rakuten).

Change-Id: I96e56dff61238e3c12cc34c643e8b3dbfeac6c11
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5294418
Commit-Queue: Hugo Benichi <hugobenichi@google.com>
Tested-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Aleksander Morgado <aleksandermj@google.com>
Reviewed-by: Taoyu Li <taoyl@chromium.org>