shill: cellular: Add comment to clarify early return on ConfigureAttachApn
Add a clarifying comment on the early return in ConfigureAttachApn. The
early return cancels the user's request to configure the attach APN when
the modem is not yet in Enabled state, but this has no effect since shill
will trigger the same attach logic as soon as the modem reaches the
enabled state.

BUG=None
TEST=None

Change-Id: I402e7e4d72adc9f63c4ea22013cf32eb1561cfd6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5307998
Commit-Queue: Andrew Lassalle <andrewlassalle@chromium.org>
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>