cryptohome: Fix a problem where too-large salt files get loaded
When support for two salt locations was added, this introduced a
regression where if you have a legacy salt that's too large cryptohome
will try and load it instead of treating it as an error. If the file is
large enough this will trigger an OOM, although even if it's not that
large it still shouldn't be loaded.

This change also includes some fixes to the fuzzer to properly
initialize the UDA object it uses, because this wasn't really being done
properly (and is itself probably a regression in the fuzzer, albeit one
masked by the salt bug).

This rewrites the failure handling significantly. Now, if any kind of
failure occurs when attempting to read the legacy path, it will always
try to access the new path instead of erroring immediately. It will also
attempt to remove the existing bad file at the legacy path.

BUG=b:320731554
TEST=FEATURES=test emerge-$BOARD cryptohome
TEST=ran fuzzer with testcase from the attached bug

Change-Id: Idfa38fc3808daf724252500d82e00f6cd76f9fb4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5297708
Reviewed-by: Daniil Lunev <dlunev@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>