shill: cellular: Don't allow concurrent attach requests
Shill does not prevent 2 concurrent requests to set the initial EPS
bearer settings. When the request to set the attach APN comes from the
user, restart the round robin when the attach dbus call completes; and
when the request comes from shill itself, as in the case of an update in
the modem profiles, or a deferred request, only restart the round robin
if the modem hasn't registered.

BUG=b:324637037
TEST=tast run limozeen cellular.ShillCustomApn.*
tast run crota cellular.ShillCustomApn.*
tast run redrix cellular.ShillCustomApn.*

Change-Id: Iccc4b04199023712e2be8c983590427bd2cee80a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5308000
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>
Commit-Queue: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Aleksander Morgado <aleksandermj@google.com>