shill: set service online when the portal detection is disabled
Originally, when the portal detection is disabled, the Service stops the
network validation only and rely on the callback to set the service
state to online. However, if the NetworkMonitor is not running, the
Network doesn't send the callback. In this situation, the service state
won't be set to online correctly.

This CL fixes this issue by setting the state first, then stop the
network validation.

BUG=b:324345075
TEST=FEATURES=test emerge-brya shill
TEST=tast run $(DUT) network.ShillCaptivePortalDNS

Change-Id: Ib5de3cb4f7f23e57f8f2e006834bbf24346caa08
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5305800
Commit-Queue: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Tested-by: Chih-Yu Huang <akahuang@chromium.org>