shill: cellular: Decrease the wait to attach on unknown carriers
Currently, we reuse the timeout kTimeoutSetNextAttachApn when scheduling
an attach on an unknown carrier. There is no need to wait for that
long(~13 seconds), as identifying the carrier should be very quick, and
after a short period of time we can assume we already have all the
information to decide if the carrier is known or not. A good compromise
seems to be 3 seconds, which is long enough to get all the SIM info,
but not too long for the user to notice a disconnect if the modem
dettaches and reattaches to the network.

BUG=b:325318968
TEST=Use the modb override callbox_unknown_carrier.pbf. restart shill
and verify that the attach occurs 3 seconds after the modem is enabled.

Change-Id: I947d8f1a16a825a34d92a04a02803eac06453562
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5307999
Reviewed-by: Aleksander Morgado <aleksandermj@google.com>
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>
Commit-Queue: Andrew Lassalle <andrewlassalle@chromium.org>