shill: cellular: Do not allow a connect while registering
Do not allow connection attempts when shill is executing the command
SetInitialEpsBearerSettings. This is because ModemManager will accept
the command and both commands will run in parallel with erratic
behavior.

BUG=b:324580941
TEST=tast run crota cellular.ShillCustomApn.set_user_apn_list
Verify in the logs that when the test tries to connect inmediately after
updating the attach APN, shill delays the connection attempt until after
the attach attempt is completed.

Change-Id: I1c25af84f7e7dc7279596e7c1ae3d33ee6e94794
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5308001
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>
Commit-Queue: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Aleksander Morgado <aleksandermj@google.com>