patchpanel: Do not reset routing tag in OUTPUT chain
This patch replaces the iptables rule:
  -A OUTPUT -j CONNMARK --restore-mark
     --nfmask 0xffff0000 --ctmask 0xffff0000
with
  -A OUTPUT -m mark --mark 0x0/0xffff0000 -j CONNMARK --restore-mark
     --nfmask 0xffff0000 --ctmask 0xffff0000

When the socket has configured with a fwmark which has a routing tag,
its egress packets will contain this routing tag, but it will get reset
by the first rule for the first packet, which may cause some unexpected
routing issue.

Note that this won't be a problem without VPN, since the source ip
selection is still correct, and no routing tag will be written to the
packet before routing decision, so the packet can still be matched with
the oif or src ip rule in the rule table. But when there is a VPN, if
the original routing tag is pointing to a physical interface, it will be
reset and then overwritten with the vpn routing tag in the
apply_vpn_mark chain if this is user traffic, and thus the packet will
be routed to the vpn interface with the wrong src ip.

BUG=b:322083502
TEST=FEATURES="test" emerge-$BOARD patchpanel
TEST=tast run $DUT network.RoutingTagSocket (still WIP locally)

Change-Id: I59f32b3d8559d2ea0948690a62a56ceac5a2aa95
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5306875
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>