patchpanel: Install ip rule for ROUTE_ON_VPN mark
Install the following rule for both IPv4 and IPv6:
  32763: from all fwmark 0x8000/0xc000 lookup 1010

Similar to the current uid rules, this rule is mainly for src ip
selection. Without this rule, a packet belonging to a system process
with the ROUTE_ON_VPN mark will have the src ip on the default physical
network, but be routed to the VPN interface.

BUG=b:322083502
TEST=FEATURES="test" emerge-$BOARD patchpanel
TEST=tast run $DUT network.RoutingTagSocket (still WIP locally)

Change-Id: I6f7beae44236e45336b100f2d034d06193d085a5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5306876
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>