common-mk: libstorage: Fix fake_platform shared lib
libstorage_fake_platform should depend on libstorage_platform so it can
access the symbols from it. Otherwise packages depending on
libstorage_fake_platform will fail to find some symbols.

This results in 2 other necessary changes:
The first is -rpath=$ORIGIN ldflag needs to be added because
libstorage_fake_platform.so needs to find libstorage_platform.so in the
same directory.
The second is that we should not link libstorage_platform_static and
libstorage_platform together, so we need to build a
libstorage_fake_platform_static for libstorage unittests to depend on as
well.

BUG=b:325996080
TEST=libstorage unittests
TEST=attestation unittests in https://crrev.com/c/5300623

Change-Id: Ic3d300a8b4053caec19b85f9c466e3d47669b3b1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5308562
Reviewed-by: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Sarthak Kukreti <sarthakkukreti@google.com>
Reviewed-by: Lee Presland <zland@google.com>
Reviewed-by: Yi Chou <yich@google.com>
Commit-Queue: ChromeOS Auto Retry <chromeos-auto-retry@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>
Tested-by: Howard Yang <hcyang@google.com>