power: suspend_stress_test: Show not enter PC10 case debug info
Currently, we only know not enter PC10 in previous 'freeze' attempt. To
efficiently debug this category, we also need to know all
Package C-States status. With this info, we could better triage issue.
Ex:
  When system is stuck in PC2, it can be caused by short LTR from devices.
  When system is stuck in PC8, it can be caused by display.

Example log:
  Suspend     1 of 1: sleep for  7 seconds...rtc wakealarm: 1708422417
  Suspend failed, pc10 count did not increment from 0
  Package C-States before :
  Package C2 : 1576765
  Package C3 : 6621
  Package C6 : 3221133
  Package C7 : 0
  Package C8 : 0
  Package C9 : 0
  Package C10 : 0
  Package C-States Now :
  Package C2 : 9209145
  Package C3 : 6621
  Package C6 : 8722461
  Package C7 : 0
  Package C8 : 0
  Package C9 : 0
  Package C10 : 0

With this debug info, we could easily know stuck in PC2, PC6 or PC8 are
different issues. It could be useful when we want to know the fail rate,
verify the solution, triage if there is other s0ix problems.

BUG=b:303771771
TEST=suspend_stress_test with kernel cmdline "intel_idle.max_cstate=2"

Change-Id: Id2372d82d01d1a2e1832591ad26812748807e26f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5308571
Commit-Queue: Curtis Chen <curtis.chen@intel.com>
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>
Tested-by: Curtis Chen <curtis.chen@intel.com>