Remove CallDBusMethod from brillo.
That method is inefficient/unsafe, so remove it.
The only caller is featured. Unfortunately, it is not trivial
to get rid of the case now, and rather it will need to redesign
the featured APIs.
So, instead, this CL moves the logic into featured's internal
implementation with TODO to get rid of it eventually, so that
no new usage will be expected.

BUG=b:289932268
TEST=Run unittest locally. Tryjob.

Change-Id: I2ddb22ee6c85637344a2db04a64964a7d4d91af4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5310694
Tested-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Miriam Zimmerman <mutexlox@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>
Commit-Queue: Hidehiko Abe <hidehiko@chromium.org>