cryptohome: Fix modernize-use-equals-default tidy errors
This updates a bunch of code to lean more heavily on =default for
trivial and default constructors and destructors.

In most cases I got rid of unnecessary definition-in-cc-file setups, and
used =default in the header. In a few cases I just got rid of the
declaration+definition entirely because they were redundant. In at least
one case I kept the definition in the cc file but used =default for it.
Overall fairly trivial but it modernizes and makes the codebase more
consistent.

BUG=b:326084107
TEST=unit tests
TEST=clang-tidy

Change-Id: Iff6aed4cc3683c2c090bc0083ab3d3d24acbc5e2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5311231
Reviewed-by: Zi Lin <lziest@google.com>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>