lorgnette: Make sure cached, canonical scanners are restored correctly
When an IPPUSB scanner was being restored from the cache, the set of
canonical scanners was not being updated properly.  This caused problems
when checking to see if a second scanner is a duplicate of a canonical
scanner.  Update the IPPUSB probing such that if it returns an item from
the cache, it will also update the canonical scanners list.

BUG=b:325519602
TEST=FEATURES=test emerge lorgnette
TEST=Manual test with scanner in lab

Change-Id: Ie28a5c72b340d163daae00800ec14af3975e73d9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5310147
Reviewed-by: Benjamin Gordon <bmgordon@chromium.org>
Commit-Queue: Nathan Muggli <nmuggli@google.com>
Tested-by: Nathan Muggli <nmuggli@google.com>