ml: Temporarily allow deprecated libsoda API
This proto field is marked deprecated in the next version of libsoda.
The replacement API is also made available in the next version.

This CL temporarily allows use of the deprecated API so that the libsoda
uprev can proceed without needing to synchronize with a platform2 CL.

BUG=b:277272187
TEST=CQ only

Change-Id: I5190cd79897702dfe942dfcd0e56dc4ba84045a6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5312238
Reviewed-by: Xinglong Luan <alanlxl@chromium.org>
Commit-Queue: Andrew Moylan <amoylan@chromium.org>
Tested-by: Andrew Moylan <amoylan@chromium.org>