patchpanel: Install ip rule for BYPASS_VPN mark
Install the following rule for both IPv4 and IPv6:
  32762: from all fwmark 0x4000/0xc000 lookup <default-physical-table>

This rule will be before the vpn uid rules and after the CLAT rules.

Similar to vpn uid rules, the main purpose of this rule is for src ip
selection for BYPASS_VPN mark. Without this rule, a packet from a user
socket with BYPASS_VPN will be matched by the vpn uid rules at the src
ip selection stage, which is unexpected.

BUG=b:322083502
TEST=FEATURES="test" emerge-$BOARD patchpanel
TEST=tast run $DUT network.RoutingTagSocket

Change-Id: I2a1c9ac022d101c6d94f7e278c8a35be59f8e52d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5313148
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>