vm_tools: syslog: replace --log-destination=FOO with --syslog
vmlog_forwarder currently has `--log-dstination` option which looks like
allowing you to give a path to your log location. However, actually this
option does not honor the given path, but just compares it with
pre-defined constant just to turn on and off a flag to control which
logs of syslog and per-VM logs you want to write to.

This CL fixes this behavior by replacing that confusing option with a
bool flag of `--syslog`.

In addition, it changes the behavior of the option. When `--syslog` is
true, now vmlog_forwarder copies all VM logs to syslog without disabling
per-VM log files. There's no use case today where we want to disable
per-VM logs, so this straightforward option woule be more useful to copy
log to `/var/log/messages` on the host.

BUG=b:325362584
TEST=cros_workon_make --board=brya vm_host_tools --test
TEST=boot crostini and logs are not written to syslog

Change-Id: I72a539c67620018e036571d73384edd5e49396f0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5311377
Tested-by: Takaya Saeki <takayas@chromium.org>
Reviewed-by: Junichi Uekawa <uekawa@chromium.org>
Commit-Queue: Takaya Saeki <takayas@chromium.org>