patchpanel: Avoid reseting vpn policy in apply_local_source_mark
Currently, apply_local_source_mark chain updates the local source part
and vpn policy part bits together. e.g., the rule for detecting chronos
traffic:
  -m owner --uid-owner chronos -j MARK --set-mark 0x8100/0xff00
0x8000 sets the ROUTE_ON_VPN and 0x100 sets chrome traffic source.

The issue here is that, this group of rules will always overwrite the
vpn policy, without respecting the original value of it. When a packet
from a user socket configured with BYPASS_VPN, it will be reset to
ROUTE_ON_VPN and get routed to the vpn interface.

To solve this issue, this patch split this group of rules into 3 parts:
1. match and set source which should be routed on vpn with mask 0x3f00;
2. if the packet does not have vpn policy and has source, update vpn
   policy to ROUTE_ON_VPN;
3. match and set source which should NOT be route on vpn with mask
   0x3f00.

BUG=b:190683881
BUG=b:322083502
TEST=FEATURES="test" emerge-$BOARD patchpanel
TEST=tast run $DUT network.RoutingTagSocketAPI network.VPNRouting*

Change-Id: I8d41d523cc766834bff40fa1e909f712af0b58b0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5313149
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>